refactor: Moved packing functions to utilities to be used with other code
